Organize route method types into their own file
